+++ apps/opentelemetry/src/otel_span_ets.erl
@@ -48,12 +48,12 @@ start_link(Opts) ->
 
 %% @doc Start a span and insert into the active span ets table.
 -spec start_span(otel_ctx:t(), opentelemetry:span_name(), otel_sampler:t(), otel_id_generator:t(),
-                 otel_span:start_opts(), fun(), otel_tracer_server:instrumentation_library())
+                 otel_span:start_opts(), fun(), otel_tracer_server:instrumentation_scope())
                 -> opentelemetry:span_ctx().
-start_span(Ctx, Name, Sampler, IdGeneratorModule, Opts, Processors, InstrumentationLibrary) ->
+start_span(Ctx, Name, Sampler, IdGeneratorModule, Opts, Processors, InstrumentationScope) ->
     case otel_span_utils:start_span(Ctx, Name, Sampler, IdGeneratorModule, Opts) of
         {SpanCtx=#span_ctx{is_recording=true}, Span=#span{}} ->
-            Span1 = Span#span{instrumentation_library=InstrumentationLibrary},
+            Span1 = Span#span{instrumentation_scope=InstrumentationScope},
             Span2 = Processors(Ctx, Span1),
             _ = storage_insert(Span2),
             SpanCtx;