+++ CHANGELOG.md
@@ -7,6 +7,45 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### [API]
+
+#### Added
+
+- [Span context now set in logger metadata when context is updated in process
+  dictionary](https://github.com/open-telemetry/opentelemetry-erlang/pull/394)
+- [Instrumentation Scope replaces Instrumentation
+  Library](https://github.com/open-telemetry/opentelemetry-erlang/pull/405) --
+  If you were using the record directly, please use the function
+  `opentelemetry:instrumentation_scope/3` or
+  `opentelemetry:instrumentation_library/3` to create an `instrumentation_scope`
+  record. 
+
+### [SDK]
+
+#### Added
+
+- [Instrumentation Scope replaces Instrumentation
+  Library](https://github.com/open-telemetry/opentelemetry-erlang/pull/405) --
+  If you were using the record directly, please use the function
+  `opentelemetry:instrumentation_scope/3` or
+  `opentelemetry:instrumentation_library/3` to create an `instrumentation_scope`
+  record.
+
+### Fixed
+
+- [Allow custom text propagator to be configured via application env](https://github.com/open-telemetry/opentelemetry-erlang/pull/408)
+
+### [Exporter]
+
+#### Added
+
+- [Instrumentation Scope replaces Instrumentation
+  Library](https://github.com/open-telemetry/opentelemetry-erlang/pull/405) --
+  If you were using the record directly, please use the function
+  `opentelemetry:instrumentation_scope/3` or
+  `opentelemetry:instrumentation_library/3` to create an `instrumentation_scope`
+  record.
+
 ## SDK 1.0.5 - 2022-05-20
 
 ### Fixed